# docs/swagger.yaml
basePath: /
definitions:
...
  models.Column:
    properties:
      board_id:
        type: string
      id:
        type: string
      name:
        description: todo, in progress, completed
        type: string
    type: object
...
  models.PaginatedTasks:
    properties:
      data:
        items:
          $ref: '#/definitions/models.TaskResponse'
        type: array
      pagination:
        $ref: '#/definitions/models.Pagination'
    type: object
  models.Pagination:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  models.Task:
    properties:
      board_id:
        type: string
      category:
        type: string
      column_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      destination_id:
        type: string
      due_date:
        type: string
      id:
        type: string
      priority:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.TaskResponse:
    properties:
      board_id:
        type: string
      category:
        type: string
      column:
        allOf:
        - $ref: '#/definitions/models.Column'
        description: populated via $lookup
      created_at:
        type: string
      description:
        type: string
      destination_id:
        type: string
      due_date:
        type: string
      id:
        type: string
      priority:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateColumn:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  models.UpdateTask:
    properties:
      board_id:
        description: client sends this
        type: string
      category:
        type: string
      column_id:
        type: string
      description:
        type: string
      due_date:
        type: string
      priority:
        type: string
      title:
        type: string
    type: object
...
  utils.AppError:
    properties:
      code:
        description: HTTP status code
        type: integer
      message:
        description: error message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple task tracker API built with Go and MongoDB
  title: Task Tracker API
  version: "1.0"
paths:
...
  /check/{id}:
    get:
      description: Validate a numeric ID from the URL path
      parameters:
      - description: Numeric ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - BearerAuth: []
      summary: Check ID
      tags:
      - utils
  /columns/{id}:
    patch:
      consumes:
      - application/json
      description: Update the name of a column. Valid names are todo, in progress,
        completed. All tasks in the column automatically reflect the new name via
        $lookup.
      parameters:
      - description: Column ID
        in: path
        name: id
        required: true
        type: string
      - description: Column update object
        in: body
        name: column
        required: true
        schema:
          $ref: '#/definitions/models.UpdateColumn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - BearerAuth: []
      summary: Update a column
      tags:
      - columns
...
  /tasks:
    get:
      description: Get paginated tasks for a board with optional filters
      parameters:
      - description: Board ID (required)
        in: query
        name: board_id
        type: string
      - description: Filter by column ID
        in: query
        name: column_id
        type: string
      - description: Filter by assignee ID
        in: query
        name: destination_id
        type: string
      - description: Filter by priority (low, medium, high)
        in: query
        name: priority
        type: string
      - description: Search in title and description
        in: query
        name: search
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedTasks'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task on a board. A default todo column is created
        automatically if one does not exist for the board.
      parameters:
      - description: Task object
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AppError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - BearerAuth: []
      summary: Create a task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Permanently delete a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - tasks
    get:
      description: Get a single task by ID with column details populated
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - BearerAuth: []
      summary: Get a task
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Update one or more fields of an existing task. Only provided fields
        are updated.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - tasks
  /tasks/{id}/complete:
    patch:
      consumes:
      - application/json
      description: Move a task to a different column
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Column object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Column'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - BearerAuth: []
      summary: Update task column
      tags:
      - tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"